requirements.txt
geopy==2.2.0
python_telegram_bot==13.11
requests==2.27.
